requirements.txt
certifi>=2023.7.22
charset-normalizer==3.1.0
colorama==0.4.6
customtkinter==5.1.0
darkdetect==0.7.2
diffusers==0.19.0
filelock==3.12.0
huggingface-hub==0.15.0
idna==3.4
importlib-metadata==6.8.0
numpy==1.24.3
packaging==23.0
Pillow==9.5.0
pyparsing==3.0.9
PyYAML==6.0
regex==2023.5.5
requests==2.31.0
tk==0.1.0
tokenizers==0.13.3
torch==2.1.1
torchaudio==2.1.1
torchvision==0.15.1
tqdm==4.65.0
transformers==4.29.0
typing_extensions==4.7.1
urllib3==1.26.12
zipp==3.15.0
